Fixed automap rendering after COAL integration.

+++ src/am_map.h
@@ -60,7 +60,7 @@ void AM_Ticker(void);
 
 // Called by main loop,
 // called instead of view drawer if automap active.
-void AM_Drawer(int x, int y, int w, int h, mobj_t *focus);
+void AM_Drawer(float x, float y, float w, float h, mobj_t *focus);
 
 // Called to force the automap to quit
 // if the level is completed while it is up.
